refactor to a shared component
